+++ ViennaNET.Logging/Log4NetImpl/CustomRollingFileAppender.cs
@@ -100,7 +100,7 @@ namespace ViennaNET.Logging.Log4NetImpl
     /// Used by the internal logger to record the Type of the
     /// log message.
     /// </remarks>
-    private static readonly Type declaringType = typeof(CustomRollingFileAppender);
+    private readonly static Type declaringType = typeof(CustomRollingFileAppender);
 
 
     /// <summary>
@@ -111,8 +111,6 @@ namespace ViennaNET.Logging.Log4NetImpl
     #endregion
 
     #region Private Instance Fields
-
-    private readonly object _lockTarget = new object();
 
     /// <summary>
     /// This object supplies the current date/time.  Allows test code to plug in
@@ -598,7 +596,7 @@ namespace ViennaNET.Logging.Log4NetImpl
     /// </remarks>
     protected override void OpenFile(string fileName, bool append)
     {
-      lock (_lockTarget)
+      lock (this)
       {
         fileName = GetNextOutputFileName(fileName);